--- backend/jest.setup.js
@@ -1,1 +1,0 @@
-require('jest-fetch-mock').enableMocks()
