+++ client/photoAlbum.js
