--- jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    testEnvironment: 'jest-environment-jsdom'
-  }; // srodowisko, w ktorym będą wykonywane testy, ma używać JSDOM (pozwala na symulację przeglądarki i manipulację DOM w testach.)
-
