+++ jest/jest.setup.js
@@ -1,0 +1,1 @@
+const apiURL = 'https://jsonplaceholder.typicode.com';
