+++ jest/package.json
@@ -19,6 +19,7 @@
     "supertest": "^6.3.4"
   },
   "dependencies": {
+    "axios": "^1.6.8",
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "node-fetch": "^2.7.0"
